backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python-headless==4.10.0.84
numpy<2.0
pillow>=10.0.0
